adding yum repo flag for Redhat
In environments where RHEL uses Satellite or production servers can't
reach out to the internet, and the Yum repo is not wanted. This commit
adds a flag to disable the functionality. It also removes the
enablerepo: nginx line from the package since a) it's redundant and b)
working around it overly complicates  the situation.